fix(offline): compact cover sync row
Collapse the cover sync controls into a single tasks-page row while keeping the library picker sheet and inline progress visible during sync.
