+++ KMReader/Features/Offline/Views/OfflineCoverSyncProgressView.swift
@@ -9,32 +9,31 @@ struct OfflineCoverSyncProgressView: View {
   let progress: OfflineCoverSyncProgress
 
   var body: some View {
-    VStack(alignment: .leading, spacing: 4) {
+    HStack(spacing: 8) {
       ProgressView(value: progress.progressFraction)
+        .frame(width: 56)
         .animation(.easeInOut(duration: 0.2), value: progress.progressFraction)
-      HStack {
-        Text(
-          String(
-            localized:
-              "offline.coverSync.progressChecked \(progress.checkedCount) \(progress.totalCount)"
-          )
+      Text(
+        String(
+          localized:
+            "offline.coverSync.progressChecked \(progress.checkedCount) \(progress.totalCount)"
         )
+      )
+      .contentTransition(.numericText())
+      .animation(.default, value: progress.checkedCount)
+      .animation(.default, value: progress.totalCount)
+      Text(String(localized: "offline.coverSync.progressSynced \(progress.storedCount)"))
         .contentTransition(.numericText())
-        .animation(.default, value: progress.checkedCount)
-        .animation(.default, value: progress.totalCount)
-        Spacer()
-        Text(String(localized: "offline.coverSync.progressSynced \(progress.storedCount)"))
+        .animation(.default, value: progress.storedCount)
+      if progress.failedCount > 0 {
+        Text(String(localized: "offline.coverSync.progressFailed \(progress.failedCount)"))
           .contentTransition(.numericText())
-          .animation(.default, value: progress.storedCount)
-        if progress.failedCount > 0 {
-          Text(String(localized: "offline.coverSync.progressFailed \(progress.failedCount)"))
-            .contentTransition(.numericText())
-            .animation(.default, value: progress.failedCount)
-        }
+          .animation(.default, value: progress.failedCount)
       }
-      .font(.caption2)
-      .foregroundColor(.secondary)
-      .monospacedDigit()
     }
+    .font(.caption2)
+    .foregroundStyle(.secondary)
+    .lineLimit(1)
+    .monospacedDigit()
   }
 }
